requirements.txt
certifi==2017.7.27.1
Django==1.11.7
django-vanilla-views==1.0.4
oauthlib==2.0.6
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
traitlets==4.3.2